Split orc deserializers into its own file.
